requirements.txt

python-telegram-bot==20.7
stripe>=9.8
python-dotenv>=1.0
SQLAlchemy>=2.0
Flask>=3.0
waitress>=3.0
APScheduler>=3.10
Flask-Admin>=1.6
